@@ -1,1 +1,1 @@
-ruby-2.6.5
+ruby-3.2.2
