@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2024_08_20_115735) do
+ActiveRecord::Schema.define(version: 2024_08_21_100558) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -90,7 +90,17 @@ ActiveRecord::Schema.define(version: 2024_08_20_115735) do
     t.integer "baseappfeature_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.text "description"
+  end
+
+  create_table "otps", force: :cascade do |t|
+    t.integer "user_id"
+    t.string "otp_number"
+    t.datetime "otp_expiry"
+    t.string "type"
+    t.string "email"
+    t.string "mobile_number"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
   end
 
   create_table "user_app_documents", force: :cascade do |t|
@@ -158,6 +168,11 @@ ActiveRecord::Schema.define(version: 2024_08_20_115735) do
     t.datetime "updated_at", precision: 6, null: false
     t.string "plaintext_password"
     t.string "profile"
+    t.boolean "email_verified"
+    t.boolean "mobile_verification"
+    t.string "type"
+    t.string "role"
+    t.boolean "accepted_terms_conditions"
   end
 
 end
